requirements.txt
google-genai
openai
python-dotenv
firecrawl
tiktoken
rich